김민승 : 캐릭터 피격 초반부

+++ Source/Sevendays_to_survive/Player/Global/DataTable/C_PlayerDataTable.h
@@ -33,7 +33,7 @@ public:
 	int staminaCalValue=0;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Resources")
-	float Hp=0.0f;
+	int Hp=0.0f;
 
 };
